Cambios dto mapper storage
Comienzo de dto, mapper y storage
