+++ entregas/Gregorio_Asensio_Christian/src/main/kotlin/services/storage/base/Storage.kt
@@ -1,4 +1,6 @@
 package org.example.services.storage.base
 
-interface Storage {
+interface Storage<T> {
+    fun store(data: List<T>, fileName: String): Boolean
+    fun load(fileName: String): List<T>
 }
