+++ entregas/Gregorio_Asensio_Christian/src/main/kotlin/services/storage/base/tenistas/StorageTenistasCsv.kt
@@ -1,0 +1,3 @@
+package org.example.services.storage.base.tenistas
+
+class StorageTenistasCsv
